Move feature vector from ExecutionTrace to ExecutionResult
This was necessary since the trace is not available in the execution observer
