+++ client/src/main/java/org/evosuite/ga/metaheuristics/mapelites/TestResultObserver.java
@@ -66,7 +66,7 @@ class TestResultObserver extends ExecutionObserver implements Serializable {
   public void testExecutionFinished(ExecutionResult result, Scope scope) {
     for(Object instance : scope.getObjects(this.targetClass)) {
       FeatureVector vector = new FeatureVector(this.inspectors, instance);
-      result.getTrace().addFeatureVector(vector);
+      result.addFeatureVector(vector);
     }
   }
